Refactor.
Separate server into specialized concerns inside a package. The
application is run via server.py which initiates the 'liberated' module.
Inside 'liberated,' '__init__' automatically executes. It bootstraps the
flask application, connects to the database, and imports the route
responses. This is in 'responses.py.'
