--- app/__init__.py
@@ -1,4 +1,0 @@
-from flask import Flask
-# from make_json_app import make_json_app
-# app = make_json_app(__name__)
-import app.responses
