--- app/responses.py
@@ -1,7 +1,0 @@
-from flask import render_template
-from app import app
-
-@app.route('/')
-def hello_world():
-    return render_template('index.html')
-
