+++ liberated/responses.py
@@ -1,0 +1,100 @@
+from flask import request, render_template, g, redirect, Response
+from liberated import app
+
+#
+# @app.route is a decorator around index() that means:
+#   run index() whenever the user tries to access the '/' path using a GET request
+#
+# If you wanted the user to go to, for example, localhost:4111/foobar/ with POST or GET then you could use:
+#
+#       @app.route('/foobar/', methods=['POST', 'GET'])
+#
+# PROTIP: (the trailing / in the path is important)
+# 
+# see for routing: http://flask.pocoo.org/docs/0.10/quickstart/#routing
+# see for decorators: http://simeonfranklin.com/blog/2012/jul/1/python-decorators-in-12-steps/
+#
+@app.route('/')
+def index():
+    """
+    request is a special object that Flask provides to access web request information:
+
+    request.method:   'GET' or 'POST'
+    request.form:     if the browser submitted a form, this contains the data in the form
+    request.args:     dictionary of URL arguments, e.g., {a:1, b:2} for http://localhost?a=1&b=2
+
+    See its API: http://flask.pocoo.org/docs/0.10/api/#incoming-request-data
+    """
+
+    # DEBUG: this is debugging code to see what request looks like
+    # print request.args
+
+    #
+    # example of a database query
+    #
+    cursor = g.conn.execute('SELECT name FROM test')
+    names = []
+    for result in cursor:
+        names.append(result['name'])  # can also be accessed using result[0]
+    cursor.close()
+
+    #
+    # Flask uses Jinja templates, which is an extension to HTML where you can
+    # pass data to a template and dynamically generate HTML based on the data
+    # (you can think of it as simple PHP)
+    # documentation: https://realpython.com/blog/python/primer-on-jinja-templating/
+    #
+    # You can see an example template in templates/index.html
+    #
+    # context are the variables that are passed to the template.
+    # for example, 'data' key in the context variable defined below will be
+    # accessible as a variable in index.html:
+    #
+    #     # will print: [u'grace hopper', u'alan turing', u'ada lovelace']
+    #     <div>{{data}}</div>
+    #     
+    #     # creates a <div> tag for each element in data
+    #     # will print: 
+    #     #
+    #     #   <div>grace hopper</div>
+    #     #   <div>alan turing</div>
+    #     #   <div>ada lovelace</div>
+    #     #
+    #     {% for n in data %}
+    #     <div>{{n}}</div>
+    #     {% endfor %}
+    #
+    context = dict(data = names)
+
+    #
+    # render_template looks in the templates/ folder for files.
+    # for example, the below file reads template/index.html
+    #
+    return render_template('index.html', **context)
+
+#
+# This is an example of a different path.  You can see it at:
+# 
+#     localhost:4111/another
+#
+# Notice that the function name is another() rather than index()
+# The functions for each app.route need to have different names
+#
+@app.route('/another')
+def another():
+    return render_template('another.html')
+
+
+# Example of adding new data to the database
+@app.route('/add', methods=['POST'])
+def add():
+    name = request.form['name']
+    g.conn.execute('INSERT INTO test VALUES (NULL, ?)', name)
+    return redirect('/')
+
+
+@app.route('/login')
+def login():
+    abort(401)
+    this_is_never_executed()
+
