+++ liberated/templates/index.html
@@ -17,7 +17,7 @@
     {% endfor %}
   </div>
 
-<p><a href="another.html">Go to the other page</a></p>
+<p><a href="another">Go to the other page</a></p>
 
 <form method="POST" action="/add">
 <p>Name of new computer scientist: <input type="text" name="name"> <input type="submit" value="Add"></p>
