+++ requirements.txt
@@ -1,4 +1,5 @@
 Flask==0.10.1
+PyYAML==3.10
 SQLAlchemy==1.0.12
-psycopg2==2.6.1
 click==6.4
+psycopg2==2.6.1
