--- run.py
@@ -1,2 +1,0 @@
-from app import app
-app.run(debug=True, host="0.0.0.0")
